requirements.txt
absl-py @ file:///croot/absl-py_1686852429912/work
accelerate==1.1.1
aif360==0.6.1
aiohttp @ file:///croot/aiohttp_1701112538292/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annexremote==1.6.6
astropy==6.1.2
astropy-iers-data==0.2024.8.12.0.32.58
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
axial_positional_embedding==0.3.12
backports.tarfile==1.2.0
boto3==1.35.22
botocore==1.35.22
braindecode==0.8.1
bypy==1.8.9
chardet==5.2.0
CoLT5-attention==0.11.1
contourpy==1.2.0
coverage==7.5.1
cycler==0.12.1
datalad==1.1.3
decorator==4.4.2
dill==0.4.0
distro==1.9.0
dlib==19.24.2
dm-tree==0.1.8
dn3==1.2.1
docstring-inheritance==2.2.0
EDFlib-Python==1.0.8
efficientnet-pytorch==0.7.1
einops==0.8.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fairlearn==0.12.0
fasteners==0.19
fonttools==4.45.1
fsspec==2023.10.0
future==1.0.0
h5py==3.10.0
hdf5storage==0.1.19
hrv-analysis==1.0.4
huggingface-hub==0.26.3
humanize==4.10.0
hyper-connections==0.1.15
imageio==2.33.0
imageio-ffmpeg==0.4.9
inFairness==0.2.3
iniconfig==2.0.0
inplace-abn==1.1.0
iso8601==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.15.0
keyring==25.4.0
keyrings.alt==5.0.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
linear-attention-transformer==0.19.1
linformer==0.2.3
local-attention==1.11.1
looseversion==1.3.0
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
memory-profiler==0.61.0
mkl-service==2.4.0
ml-dtypes==0.2.0
mne==0.20.3
mne-bids==0.13
moabb==1.0.0
more-itertools==10.5.0
moviepy==1.0.3
msgpack==1.1.0
mtcnn==0.1.1
multiprocess==0.70.18
namex==0.0.7
neurokit2==0.2.10
nltk==3.8.1
nolds==0.6.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.2.3
parse==1.20.2
patool==2.4.0
patsy==0.5.6
peft==0.13.2
pluggy==1.5.0
pooch==1.8.1
POT==0.9.5
product_key_memory==0.2.11
proglog==0.1.10
protobuf==3.20.3
pyasn1-modules==0.2.8
pyEDFlib==0.1.38
pyelftools==0.32
pyerfa==2.0.1.4
pyparsing==3.1.1
pyproject==1.3.1
pyriemann==0.5
pytest==7.4.4
python-gitlab==4.11.1
PyYAML==5.3.1
pyyaml-include==1.2
randaugment==1.0.2
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rich==13.7.0
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.3.2
scikit-posthocs==0.9.1
seaborn==0.12.2
SecretStorage==3.3.3
skorch==0.15.0
soundfile==0.12.1
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.11
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.1.1
torchinfo==1.8.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.46.3
triton==3.2.0
typing_extensions==4.12.2
unix-ar==0.2.1
visdom==0.2.4
websocket-client==1.6.4
wfdb==4.1.2
wget==3.2
xlrd==2.0.1
zstandard==0.23.0
